cleanup seekerApplication remove dead code
